LayoutTests/fast/html/HTMLOptionsCollection-set-property-expected.txt

Tests setting properties on HTMLOptionsCollection

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS collection.selectedIndex is 0
PASS collection.action is document.getElementById('action')
PASS collection[0] is document.getElementById('selectedIndex')
PASS collection.testOption is document.getElementById('testOption')
PASS collection.foo is undefined.
collection.selectedIndex = 1
PASS collection[0] = 'bar' threw exception TypeError: Type error.
PASS collection.selectedIndex is 1
firstOption.id = 'bar'
firstOption.name = 'bar'
PASS collection.selectedIndex is 1
collection.action = 'about:blank'
PASS collection.action is document.getElementById('action')
document.getElementById('action').remove()
PASS collection.action is undefined.
collection.testOption = 'test'
PASS collection.testOption is document.getElementById('testOption')
document.getElementById('testOption').remove()
PASS collection.testOption is undefined.
collection.foo = 'test'
PASS collection.foo is "test"
PASS collection[0] is firstOption
PASS successfullyParsed is true

TEST COMPLETE
